+++ main.go
@@ -9,13 +9,13 @@ import (
 
 	"github.com/hashicorp/packer-plugin-sdk/plugin"
 
-	"github.com/hashicorp/packer-plugin-lxd/builder/lxd"
-	"github.com/hashicorp/packer-plugin-lxd/version"
+	"github.com/bketelsen/packer-plugin-incus/builder/incus"
+	"github.com/bketelsen/packer-plugin-incus/version"
 )
 
 func main() {
 	pps := plugin.NewSet()
-	pps.RegisterBuilder(plugin.DEFAULT_NAME, new(lxd.Builder))
+	pps.RegisterBuilder(plugin.DEFAULT_NAME, new(incus.Builder))
 	pps.SetVersion(version.PluginVersion)
 	err := pps.Run()
 	if err != nil {