+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ## 1.0.0 (June 14, 2021)
 
 * Update packer-plugin-sdk to v0.2.3
-* Add support for LXD virtual-machine type images. [GH-12]
+* Add support for incus virtual-machine type images. [GH-12]
 
 ## 0.0.1 (April 22, 2021)
 
-* LXD Plugin break out from Packer core. Changes prior to break out can be found in [Packer's CHANGELOG](https://github.com/hashicorp/packer/blob/master/CHANGELOG.md).
+* incus Plugin break out from Packer core. Changes prior to break out can be found in [Packer's CHANGELOG](https://github.com/hashicorp/packer/blob/master/CHANGELOG.md).